requirements.txt
PyQt5
python-vlc
librosa
numpy
matplotlib
Pillow
